edwSQL

passes a SQL query from a separate .sql file to the EDW. Commented lines in the .sql file must be preceded by '--'. Blocked comments are not supported.

edwSQL(sql = NULL, resource = "custom", custom = NULL, file = TRUE,
  DSN = FALSE, uid = NULL, pwd = NULL, ...)

Arguments

sql
- name of a .sql file

resource
- Alias for the database connection

custom

allows for a connection to any database server. The connection string must be passed using the custom parameter.

EDWDBDev

EDWDBDev

EDWDBProd

EDWDBProd

Trantor

EDWDBProd

Phloston

EDWDBDev

ClinicalAnalytics

\\wn2591\PremierPRD

Apollo

Apollo_Spokane

PHC_CA

PHC Clinical Analytics Database

custom
full connection string when using 'custom' as the resource.

file
switch indicating if sql parameter is a .sql file or a command. Default is TRUE (requires .sql file)

DSN
if TRUE, use a DSN as specified in `resource` parameter instead of a connection string (ie. for Axis PATS: resource = "TSI_32", DSN = TRUE)

uid
username for connections if necessary. Default is NULL, which will then use the user's system credentials.

pwd
password for connections that don't use the user's system credentials. Default is NULL.

...
additional arguments. not used.

Value

data

Query results as a data frame

fields

field names

elapsed_seconds

elapsed system time to run the query

status_message

the number of records & variables in the results

Examples

# \donttest{
### aliased resource (ie. EDWDBDev or EDWDBProd)
mydata <- edwSQL(sql = "SQL/mydata_query.sql",
       resource = "EDWDBDev")$data

#> Warning: cannot open file 'SQL/mydata_query.sql': No such file or directory

#> Error in file(con, "r"): cannot open the connection

### Custom resource - any valid ODBC source. If uid & pwd are not specified, then
# the Windows (ie. a 'trusted connection' will be used).

mydata <- edwSQL(sql = "SQL/mydata_query.sql",
       resource = "custom",
       custom = "Driver=SQL Server;Server=EDWDBDev; Database=ProvidenceEpic")$data

#> Error: nanodbc/nanodbc.cpp:983: 00000: [unixODBC][Driver Manager]Data source name not found, and no default driver specified 

### User DSN - need to specify the uid & pwd.
# This requires a user or system DSN to be installed on your computer.
# The uid and pwd should not be set in the script or in the console for
# security reasons. Instead, set them in your .Rprofile file like this:
options(my_pwd = "mypassword", my_uid = "username")

# then in your script, you can load the uid and pwd like this:
uid <- getOption("my_uid")
pwd <- getOption("my_pwd")
mydata <- edwSQL(sql = "SQL/mydata_query.sql",
       resource = "ClinicalAnalytics", file = FALSE, DSN = TRUE)$data

#> Error: nanodbc/nanodbc.cpp:983: 00000: [unixODBC][FreeTDS][SQL Server]Unable to connect: Adaptive Server is unavailable or does not exist  [unixODBC][FreeTDS][SQL Server]Unable to connect to data source 

# }